+++ arch/x86/include/asm/atomic64_64.h
@@ -187,7 +187,7 @@ static inline bool arch_atomic64_add_negative(long i, atomic64_t *v)
 #ifndef CONFIG_KTSAN
 	return GEN_BINARY_RMWcc(LOCK_PREFIX "addq", v->counter, s, "er", i);
 #else
-	return (ktsan_atomic64_fetch_add((void *)v, i,
+	return ((long)ktsan_atomic64_fetch_add((void *)v, i,
 			ktsan_memory_order_acq_rel) + i) < 0;
 #endif
 }
